+++ lib/geometry.ts
@@ -1,0 +1,46 @@
+import type { Point3, Box } from "./types"
+import { add, rotLocal } from "./vec3"
+
+// faces: front, back, bottom, top, right, left
+export const FACES: [number, number, number, number][] = [
+  [0, 1, 2, 3],
+  [4, 7, 6, 5],
+  [0, 1, 5, 4],
+  [3, 2, 6, 7],
+  [1, 2, 6, 5],
+  [0, 3, 7, 4],
+]
+export const EDGES: [number, number][] = [
+  [0, 1],
+  [1, 2],
+  [2, 3],
+  [3, 0],
+  [4, 5],
+  [5, 6],
+  [6, 7],
+  [7, 4],
+  [0, 4],
+  [1, 5],
+  [2, 6],
+  [3, 7],
+]
+export const TOP = [3, 2, 6, 7] as const
+
+export function verts(b: Box): Point3[] {
+  const {
+    size: { x: sx, y: sy, z: sz },
+    center,
+    rotation,
+  } = b
+  const offs = [
+    { x: -sx / 2, y: -sy / 2, z: -sz / 2 },
+    { x: sx / 2, y: -sy / 2, z: -sz / 2 },
+    { x: sx / 2, y: sy / 2, z: -sz / 2 },
+    { x: -sx / 2, y: sy / 2, z: -sz / 2 },
+    { x: -sx / 2, y: -sy / 2, z: sz / 2 },
+    { x: sx / 2, y: -sy / 2, z: sz / 2 },
+    { x: sx / 2, y: sy / 2, z: sz / 2 },
+    { x: -sx / 2, y: sy / 2, z: sz / 2 },
+  ]
+  return offs.map((o) => add(center, rotLocal(o, rotation)))
+}
